eve: Modify 1.5 vs 3.0 current limit setting for P1B boards
P1B adds 2 new GPIO signals to explicitly control the 1.5 vs 3.0
current limit selection. In addition, the use case for the 5V enable
gpio changes to on/off and no longer needs to be toggled between input
and output.

This change is tied to the board version so that the operation for P1
and older boards is unaffected.

BUG=chrome-os-partner:61431
BRANCH=none
TEST=Manual
Connected display adapter into C0 port and verified gpio signals:
  1  USB_C0_5V_EN
  0* USB_C1_5V_EN
  1* EN_USB_C0_3A
  0  EN_USB_C1_3A
Then connect another adapter to C1:
  1  USB_C0_5V_EN
  1* USB_C1_5V_EN
  0* EN_USB_C0_3A
  0  EN_USB_C1_3A
Remove adapter from C0:
  0* USB_C0_5V_EN
  1  USB_C1_5V_EN
  0  EN_USB_C0_3A
  1* EN_USB_C1_3A

Also tested P1 system to verify that the correct board version is read
and the signals match the expected value.

Change-Id: Iebab84304a33282c766b8963dba4a56052a295c7
Signed-off-by: Scott <scollyer@chromium.org>