Update patch set 2
Patch Set 2:

> > do I get it right that the idea now is that ccd is enabled when the ccd_mode_l pin goes low, and sometimes it goes low because we drive it?
> 
> Yes, that's correct.

I am not sure if this is completely sane: what if there is a device where this pin is pulled so strongly that h1 can't even pull it down. This would be a hardware problem, but we'd have to be able to deal with it nevertheless.

Mary, what do you think of this approach?

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>