apollolake: add a slp_s0 signal detect mechanism
This adds the ability to do a SLP_S0 signal assertion
failure detection in EC. When the host command for S0ix
entry is received, EC will start a deferred timer
which will check the GPIO value after the timer expires.
If the GPIO is not asserted then, EC will wake up the system
by using forced lid open if lid is closed or by pulsing
the PCH.

BRANCH=None
BUG=None
TEST=In conditions where system does not assert the
SLP_S0 signal, check if EC is able to wake up the system.

Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>

Change-Id: I3ae6d39b560bfa3be943dbec2b069663549e9f7e