Update patch set 3
Patch Set 3: Code-Review+2

I thought I +2'd this. It's still hacky to use a global. I think it'd be better to return an enum that represents the correct state of the battery instead of hacking in this logic in the led code.

Patch-set: 3
Reviewer: Gerrit User 1002124 <1002124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2