cr50: enable utmi wakeups
At one point we disabled wakeups use of the AP phy, so we disabled UTMI
as a wake source when that phy was selected. When we reenabled use of
the AP phy, we forgot to reenable USB as a wake source when that phy was
selected. Because wakeup was disabled, after Cr50 went to sleep any usb
transactions would drop bits, because Cr50 wouldn't notice anything was
happening until it woke up on one of the HOOK_TICK events.

This change reenables UTMI wakeups on boards with AP usb. 

We haven't been configuring usb correctly for suspend. This change
writes 1 to USB_PCGCCTL_STOPCLK during init. This makes the controller
stop the phy clock whenever it detects a usb suspend. When usb resumes
out of suspend, this bit has no effect. Without this bit set the
controller doesn't tell the system it is in suspend and UTMI wakeup
events are sent immediately after entering sleep. This prevents cr50
from staying asleep.

BUG=b:35774906
BRANCH=cr50
TEST=Boot up reef. Wait until cr50 goes to sleep run 'usb_updater -f'
and verify that it runs successfully. Make sure deep sleep still works

Change-Id: I54bd866111b5c9b5738575f23757e0cbe4907ec4
Signed-off-by: Mary Ruthven <mruthven@chromium.org>