chip: npcx: Ignore LRESET/PLTRST WUI in case of eSPI
In case of eSPI, we use VW PLTRST signal to perform host register
init. Thus, LRESET/PLTRST WUI can be ignored since there is nothing to
be done. Until now this was not a problem because gpio_interrupt
cleared out all WUI bits irrespective of the fact that the gpio is valid
or not. With the recent change 9c24fac50 to fix GPIO ISR behavior,
devices usign eSPI (poppy and eve) were stuck with no handler clearing
WUI 57 for LRESET/PLTRST.

This change masks out the interrupt for LRESET/PLTRESET and clears the
interrupt pending bit so that the EC is not interrupted by this
anymore.

BUG=b:35954584
BRANCH=None
TEST=Verified that no more watchdog resets are seen on poppy.

Change-Id: Ib35b574d937c40d417ac707296ec080fd0244cf8
Signed-off-by: Furquan Shaikh <furquan@chromium.org>