cr50: disable sleep on utmi wakeup
UTMI wakeups are triggered by the USB controller on usb resume. This
change disables sleep with the usb mask. It will be reenabled when usb
suspend is detected.

Before when we woke up, we relied on Cr50 staying awake long enough for
usb_interrupt to process the interrupt, see the wakeup, and disable
sleep. This change disables sleep directly in the pmu_wakeup_interrupt
to ensure that we don't go to sleep before we can process
GR_USB_GINTSTS.

BUG=b:35774906
BRANCH=none
TEST=Run 100 'usb_updater -f' 100 times sleeping 20 seconds in between
each run. Verify there are no failures and cr50 still goes to sleep
between each run.

Change-Id: I1819deaa3988bcf2a85686d1b9d57092ba264c4d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>