cr50: change DEV to DBG in debug images version strings
There are two types of signing of CR50 images - prod and dev.
Designating images built with CR50_DEV variable set in the environment
as DEV is confusing, as this has nothing to do with the signing type
(dev vs prod), and is in fact indicating an image with many debug
features enabled.

This patch changes the string to "DBG".

BRANCH=cr50
BUG=none

TEST=verified that the modified image has correct string in the
     version field:
  > vers
  Chip:    g cr50 B2-D
  Board:   0
  RO_A:  * 0.0.10/29d77172
  RO_B:    -1.-1.-1/ffffffff
  RW_A:    0.0.16/DBG/cr50_v1.1.6137-1624610+
  RW_B:  * 0.0.16/cr50_v1.1.6137-1624610+
  Build:   0.0.16/cr50_v1.1.6137-1624610+
           tpm2:v0.0.287-1a68fe6
           cryptoc:v0.0.8-6283eee
           2017-03-06 16:51:15 vbendeb@eskimo.mtv.corp.google.com
   >

Change-Id: I06a97a6aff5418a4d02e71ca23813e6d2005da5c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>