Update patch set 6
Patch Set 6:

(5 comments)

What's the status of this CL? I stumbled across this section of code this week while I was tracking down an issue related to CCD with servo_v4. While this CL does some clean up, and fixes the bug where the contents of ANALOAG_CTRL_5 were being inadvertently written back in to ANALOG_CTRL_2, it doesn't fix the use case for servo_v4 and CCD. Namely, we need to be able to set the USB PD data role to DFP without automatically connecting the DP aux signals to sbu_1/2.

As I didn't know this CL was pending, or any of the discussion around it, I put together a CL that fixed what I thought were the problems with how the mux is being controlled. https://chromium-review.googlesource.com/c/451837/

I have added comments to this CL for the areas that I think still need to be addressed.

Patch-set: 6
CC: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>