spi: extend API to do a full readback on STM32
Extend the SPI master API to be able to read the incoming bits while
transmitting. If SPI_READBACK_ALL is set in 'rxlen' when calling
spi_transaction(), then the received data during transmission is recorded
in rxdata buffer and the function assumes that the real 'rxlen' is equal
to 'txlen'.

Implement it for STM32.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=b:35648259
TEST=on Eve FP MCU, manually do some SPI requests to the sensor.

Change-Id: Iebff617acd3230277d36a4f565766b7748721a1d