cr50: fix prod images timestamp
When sighing prod images it is important to be able to reproduce them
bit for bit. Setting the manifest timestamp filed to a non-zero value
makes sure that this value is used in the header as opposed to the
current time.

Setting the value to 1 guarantees that any dev image with the same
epoch/major/minor field values will be preferred, as it would have a
later timestamp.

BRANCH=none
BUG=none

TEST=verified that two images built with this manifest are exactly the
     same (they used to differ before this change).
     verified that the header timestamp field is indeed set to 1:

  $ od -Ax -t x1 -vd d2/cr50.bin.prod | grep -A1 004350
  $ 004350 ff ff ff ff ff ff ff ff ff ff ff ff 00 00 00 00
  $ 004360 00 00 00 00 11 00 00 00 01 00 00 00 00 00 00 00

     location 0x435c is the epoch_ field offset, 32 bit
     epoch/major/minor is set to 0/0/17 and 64 bit timestamp is set to
     1.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>

Change-Id: I6ea0e664fa3eab7917ca472d715824feec49eb51