Update patch set 3
Patch Set 3:

> Hi Vincent, why you increase the charger power to 25W? What's the specific bug about this CL? I'm considering whether to refer it for our project. Need your advice, thanks a lot!

caroline often needs more than 15W to boot + sysjump. A significant side effect of this change is that if the battery is missing / very low (and system is locked) then AP power-on will be inhibited (because we don't negotiate PD in RO, so we're stuck at 15W), so the device won't be able to boot (until battery charge is sufficient).

Patch-set: 3
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>