kevin: Increase display pwm frequency
210 is kinda low, it's visible both with a camera and while shaking the
screen or shaking a finger in front of the display.

The driver says that we can go as high as 20-30kHz, but at those
frequencies the low pwm duty cycles become very unstable. Either a black
screen or a flickering screen.

1200Hz seems like a good compromize, it's not really visible while
shaking the screen/finger. Taking a picture (with a 5x) still produces
some patterns but they're very tiny now (the moire patterns are much more
noticeable than the flickering).

BRANCH=gru
BUG=b/35648273
TEST=Kevin display still works, pwm not noticeable anymore

Change-Id: I9b245f6261ee2feef67d9021e2991cba2f3ef7fa