Update patch set 3
Patch Set 3:

> we need to disable the keyboard backlight when the lid angle is between 180 and 360 degrees

The host is the end-decider of tablet mode, so this needs to be host-controlled. I opened crbug.com/701868 to add this feature. We can add a TODO / comment in the code for now, but we shouldn't make adjustments while in S0 on the EC side.

Patch-set: 3
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>