Update patch set 3
Patch Set 3:

> The host is the end-decider of tablet mode, so this needs to be host-controlled

In fact, this should already be host-controlled based on host-determined tablet mode state. Is the host not going to tablet mode when the lid angle exceeds 180 degrees, when the tablet mode switch is not asserted? If so, we need to investigate that.

Patch-set: 3
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>