Update patch set 3
Patch Set 3:

> In fact, this should already be host-controlled based on host-determined tablet mode state. Is the host not going to tablet mode when the lid angle exceeds 180 degrees, when the tablet mode switch is not asserted? If so, we need to investigate that.

Yes, although host can disable keyboard backlight when tablet mode. we need to do this when the lid angle exceeds 180 degrees and the host is not going to tablet mode. This time the tablet mode is not asserted(use a gpio pin GPIO_TABLET_MODE_L instead of using lid angle). And I refer the cave, seems to have done this on the EC side.

Patch-set: 3