tcpm: anx74xx: Take tcpc out of suspend state before I2C access
The cable_det signal is used to signal to the EC that the tcpc has
detected a cable being connected when low power mode is
configured. The driver then needs to take the tcpc our of suspend
state by setting PWR_EN and RESET_N high. Then bit 7 of analog_ctrl_0
needs to be set properly.

The code that is handling this transition was attempting to access the
tcpc via I2C before pull putting the chip in normal mode when
transitioning from suspend to normal

BUG=b:35775019
BRANCH=reef
TEST=Put tcpc port into sink-only but running 'lidclose' on EC
console. Then connect a TypeC -> TypeA adaptor and verify that there
are no floods of "TCPC: Low Power Mode" messages on the EC
console. Also verified that can attach USB PD charger and other sink
only adapters.

Change-Id: I6e4843e43f59afbf5ca3251feb68981b815c1c78
Signed-off-by: Scott <scollyer@chromium.org>