chip/g/upgrade: Clarify pdu/frame terminology
PDU is a protocol unit that can be sent directly over TPM interface,
or that needs to be broken up and encapsulated in a _frame_ to be sent
over USB, the frame has a header that indicates its total size (since it
can be fragmented over multiple USB packets).

Also, fix an issue in update_pdu_valid, where a wrong offset was being
used (with no consequence as the wrong computation had the same resulting
value).

BRANCH=none
BUG=b:36375666
TEST=make buildall -j
TEST=cd extra/usb_updater; make
TEST=Run sudo extra/usb_updater/usb_updater cr50.bin, rebuild cr50.bin,
     and run it again, to make sure cr50 can still be updated.

Change-Id: Idb34a6eb96edbeeb74e4479a745d26ee8df572f7