common/usb_update: Fix valid_transfer_start logic
Current code calls valid_transfer_start to do 2 things: fetch
update_frame_header into a structure, and check if it looks
like an update start frame. Let's separate the 2, and just
fetch the update frame.

This used to cause a problem when trying to update RO image
for hammer, as the data is then at offset 0, and the digest
is also set to 0 (since we ignore hashes in common code), which
would make usb_updater.c believe that this was a transaction
start.

BRANCH=none
BUG=b:36375666
TEST=Can update RO from RW with usb_updater2

Change-Id: I62c1d2021530d14fe61ecc4241aee02b253f2e1a