Unlock system if battery is not present
Reef devices use battery presence pin as a WP pin. Cr50 reads the
battery pin and applies it to the WP pin. However, due to the timing
differences that cr50 toggles the WP pin and the EC reads the pin,
EC gets mixed readings.

This patch makes the EC check the battery presence PIN by itself
to determine whether the system is locked or not.

BUG=b:36659750
BRANCH=reef
TEST=Boot Electro and Snappy without battery

Change-Id: I9cb0236b73c73d4a629332abe0bc4d5e0aee52cb
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>