Delay checking WP for PD comm if battery removed
Reef devices use battery presence pin as a WP pin. Cr50 reads the
battery pin and applies it to the WP pin. However, due to the timing
differences that cr50 toggles the WP pin and the EC reads the pin,
EC gets mixed readings.

This patch makes the EC RO check the battery presence PIN. If
battery isn't present, it reads WP pin after 50 msec delay
then determines whether PD communication is allowed or not.

BUG=b:36659750
BRANCH=reef
TEST=Boot Electro and Snappy without battery. Verified PD
communication is enabled if battery is removed and disabled if
battery is present.

Change-Id: I9cb0236b73c73d4a629332abe0bc4d5e0aee52cb
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>